chore: 🔧 created a component for the sidebar info
